21 Taking Payments -- 269 Testing card failures

+++ client/src/app/checkout/checkout-payment/checkout-payment.component.ts
@@ -25,6 +25,7 @@ export class CheckoutPaymentComponent implements AfterViewInit, OnDestroy {
   cardExpiry: any;
   cardCvc: any;
   cardErrors: any;
+  cardHandler = this.onChange.bind(this);
 
   constructor(
     private basketService: BasketService,
@@ -38,12 +39,15 @@ export class CheckoutPaymentComponent implements AfterViewInit, OnDestroy {
 
     this.cardNumber = elements.create('cardNumber');
     this.cardNumber.mount(this.cardNumberElement.nativeElement);
+    this.cardNumber.addEventListener('change', this.cardHandler);
 
     this.cardExpiry = elements.create('cardExpiry');
     this.cardExpiry.mount(this.cardExpiryElement.nativeElement);
+    this.cardExpiry.addEventListener('change', this.cardHandler);
 
     this.cardCvc = elements.create('cardCvc');
     this.cardCvc.mount(this.cardCvcElement.nativeElement);
+    this.cardCvc.addEventListener('change', this.cardHandler);
   }
 
   ngOnDestroy() {
@@ -52,14 +56,36 @@ export class CheckoutPaymentComponent implements AfterViewInit, OnDestroy {
     this.cardCvc.destroy();
   }
 
+  onChange({ error }) {
+    if (error) {
+      this.cardErrors = error.message;
+    } else {
+      this.cardErrors = null;
+    }
+  }
+
   submitOrder() {
     const basket = this.basketService.getCurrentBasketValue();
     const orderToCreate = this.getOrderToCreate(basket);
     this.checkoutService.createOrder(orderToCreate).subscribe((order: IOrder) => {
       this.toastr.success('Order created successfully');
-      this.basketService.deleteLocalBasket(basket.id);
-      const navigationExtras: NavigationExtras = { state: order };
-      this.router.navigate(['checkout/success'], navigationExtras);
+      this.stripe.confirmCardPayment(basket.clientSecret, {
+        payment_method: {
+          card: this.cardNumber,
+          billing_details: {
+            name: this.checkoutForm.get('paymentForm').get('nameOnCard').value
+          }
+        }
+      }).then(result => {
+        console.log(result);
+        if (result.paymentIntent) {
+          this.basketService.deleteLocalBasket(basket.id);
+          const navigationExtras: NavigationExtras = { state: order };
+          this.router.navigate(['checkout/success'], navigationExtras);
+        } else {
+          this.toastr.error(result.error.message);
+        }
+      });
     }, error => {
       this.toastr.error(error.message);
       console.log(error);
